requirements.txt
Flask>=2.3
gunicorn
google-api-python-client
google-auth
google-cloud-secret-manager
requests
